rlatjsrnr - 연결 및 수정
