+++ mariaDB.sql
@@ -1,11 +1,14 @@
 select * from member;
-
-
 SELECT * FROM member WHERE mid = 'admin' AND mpw = 'admin';
 
-show table;
+show tables;
+select * from joinmember;
 
 select * from map;
 
+select * from wishlist;
 select * from party;
-delete from party where pnum = 5;
+
+select * from chat;
+
+select * from partyboard;
