+++ src/main/webapp/WEB-INF/views/home.jsp
@@ -1,10 +1,7 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
-<html>
-<head>
-	<title>Home</title>
-</head>
+
 <%@ include file="common/header.jsp" %>
 <%@ include file="common/search.jsp" %>
 <%@ include file="home/partyList.jsp" %>
